feat: update build process and dependencies
The build process has been updated to use glob patterns for entrypoints
instead of a single entry file. This allows for more flexibility in the
build process.

The 'bun-plugin-isolated-decl' and 'oxc-transform' dependencies have been
added. These are used to improve the build process and output.

The 'pkg-to-jsr' command now points to 'dist/bin.mjs' instead of
'dist/index.mjs'. This change reflects the new build process.

The 'bun.lockb' file has been updated to reflect the new dependencies.