feat: Move CLI entry point to separate module(`./bin`)
This commit moves the CLI entry point from 'dist/cli.js' to a separate
module under 'bin/pkg-to-jsr.mjs'. This change simplifies the build
process by eliminating the need for a shebang injection step. The
'package.json' file has been updated to reflect this change in the 'bin'
and 'scripts' sections.