feat: Add jsrInclude and jsrExclude to PackageJson type
This commit introduces two new optional fields to the PackageJson type:
jsrInclude and jsrExclude. These fields are used to customize the
inclusion and exclusion of files in the JSR configuration. The getInclude
and getExclude functions have been updated to incorporate these new
fields when generating the include and exclude lists.